Remove AlgebraicIdentifier class
and restructure responsibilities of legal_moves and algebraic_notation
